% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{cohen_d.formula}
\alias{cohen_d.formula}
\title{Cohen's d}
\usage{
\method{cohen_d}{formula}(f, data, ...)
}
\arguments{
\item{f}{A two-sided formula: y ~ x.}

\item{d}{A data frame.}

\item{test}{Whether to include a Wilcoxon (Mann-Whitney) test of location difference (default: F)}

\item{sig}{The number of significant digits to report (default: 3)}
}
\value{
Cohen's d
}
\description{
Cohen's d
}
\examples{
cohen_d(mpg ~ am, mtcars)
}
