% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{cohen_d}
\alias{cohen_d}
\title{Cohen's d}
\usage{
cohen_d(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
Cohen's d
}
\description{
Cohen's d
}
\examples{
cohen_d(x, y)
}
