% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{create_dict}
\alias{create_dict}
\title{create_dict}
\usage{
create_dict(v)
}
\arguments{
\item{v}{Character vector of names}
}
\value{
A printout out of a named vector template for a translation dictionary
}
\description{
Print out a translation dictionary given a char vector of names
}
\details{
A printout out of a named vector template for a translation dictionary
}
\examples{
\dontrun{
if(interactive()){
 create_dict(names(mtcars))
 }
}
}
