% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{custom.summarize}
\alias{custom.summarize}
\title{Output summary table split by a categorical variable}
\usage{
custom.summarize(
  df,
  vars,
  facvar = NULL,
  statscol = T,
  test_type = "wilcox",
  sig.digits = 2
)
}
\arguments{
\item{df}{data frame}

\item{vars}{Named vector. Names are variables to summarize, values are human readable versions.}

\item{facvar}{Factor variable defining exactly two groups.}

\item{statscol}{Include statistical test of location difference between two groups.}

\item{sig.digits}{The number of signficant digits for the output.}
}
\value{
A data frame with the summary stats that can be printed with kable
}
\description{
Output summary table split by a categorical variable
}
\examples{
data(mtcars)
vars = c(mpg = 'miles per gallon', hp = 'horse power')
custom.summarize(mtcars, vars, facvar='am')
}
