% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{dagitty2graph}
\alias{dagitty2graph}
\title{dagitty2graph: Graph conversion from dagitty to tbl_graph}
\usage{
dagitty2graph(g)
}
\arguments{
\item{g}{A dagitty object ("dag" or "pdag").}
}
\value{
An tbl_graph object.
}
\description{
Convert a dagitty object to a tbl_graph object.
}
\examples{
graph <- dagitty2graph(g)

}
