% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{fmt_sci}
\alias{fmt_sci}
\title{fmt_sci}
\usage{
fmt_sci(x)
}
\arguments{
\item{x}{Numeric vector to be formatted}
}
\value{
latex character vector with x in scientific notation
}
\description{
Generate latex for scientific notation
}
\details{
For use in inline markdown.
}
\examples{
\dontrun{
if(interactive()){
 fmt_sci(2.34e-9)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
