% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{forestplot}
\alias{forestplot}
\title{Create a forest plot and table of coefficients and stats from one or more regression models.}
\usage{
forestplot(
  ...,
  modelnames = NULL,
  varnames = NULL,
  stats = NULL,
  odds.ratio = F,
  breaks = NULL,
  intercept = T,
  facet = T,
  dodgewidth = 0.3,
  linewidth = 1,
  size = 1,
  draw = T,
  sig = 3
)
}
\arguments{
\item{...}{One or more model objects.}

\item{modelnames}{A character vector of names to label each plot and table column.}

\item{varnames}{A named character vector. Each element is a human friendly version of the variable name, and each name is the original variable name.}

\item{odds.ratio}{If TRUE, apply exponential transform to coefficients and omit intercept.}

\item{breaks}{An optional vector of breaks for the plot x-axis.}

\item{intercept}{If TRUE, include the intercept in the table and plot; otherwise omit.}

\item{facet}{If TRUE, display models in different facets. Otherwise, distinguish by color.}

\item{dodgewidth}{For position_dodge(). Default = 0.3.}

\item{draw}{If TRUE draw the plot; otherwise return the data. Default TRUE}

\item{sig}{The number of significant digits for the table.}
}
\value{
A ggplot object.
}
\description{
Create a forest plot and table of coefficients and stats from one or more regression models.
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ hp, mtcars)
m2 <- lm(mpg ~ wt, mtcars)
mnms <- c('MPG vs HP', 'MPG vs Weight')
vnms <- c(mpg='Miles per gallon', hp='Horsepower', wt='Weight')
forestplot(m1, m2, modelnames=mnms, varnames=vnms)
}
