% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ggdotchart}
\alias{ggdotchart}
\title{ggdotchart}
\usage{
ggdotchart(v, threshold = NULL)
}
\arguments{
\item{v}{A named vector, e.g., a table}

\item{threshold}{A numeric value. Values in v that are greater than this will have different colors and shapes}
}
\value{
a ggplot object
}
\description{
ggplot version of base::dotchart
}
\details{
Uses theme_minimal, and no axis labels.
}
\examples{
\dontrun{
if(interactive()){
 ggdotchart(table(mtcars$cyl))
 }
}
}
