% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ggecdf}
\alias{ggecdf}
\title{ggecdf}
\usage{
ggecdf(d, var, group)
}
\arguments{
\item{d}{A data frame}

\item{var}{A character string identifying the variable to plot}

\item{group}{A character string identifing the varible the point colors}
}
\value{
A ggplot
}
\description{
Plot an ECDF curve with points colored by a second variable
}
\details{
This with plot an ECDF curve of var, with each point colored by group
}
\examples{
\dontrun{
if(interactive()){
 mtcars$cyl <- factor(mtcars$cyl)
 ggecdf(mtcars, "mpg", "cyl")
 }
}
}
