% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ggmediation}
\alias{ggmediation}
\title{ggmediation}
\usage{
ggmediation(obj)
}
\arguments{
\item{obj}{An object produced by the mediate function from the mediation package.}
}
\value{
A ggplot
}
\description{
Plot ACME and ADE from objects produced by mediate function in mediation package
}
\details{
Produces a ggplot version of the mediation plot from the mediation package.
}
\examples{
\dontrun{
if(interactive()){
 library(mediation)
 m_med <- lm(hp ~ wt, mtcars)
 m_out <- lm(mpg ~ hp + wt, mtcars)
 m <- mediate(m_med, m_out, treat='wt', mediator='hp')
 ggmediation(m)
 }
}
}
