% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ggxtabs}
\alias{ggxtabs}
\title{ggxtabs}
\usage{
ggxtabs(
  xtab,
  cell_counts = F,
  stats = F,
  viridis_option = "G",
  border_color = "white"
)
}
\arguments{
\item{xtab}{xtabs object of exactly two categorical variables}

\item{cell_counts}{Display cell counts, Default: F}

\item{stats}{Display Chisq stats in caption, Default: F}

\item{viridis_option}{Viridis color palette, Default: 'G'}

\item{border_color}{Color of border between cells, Default: 'white'}
}
\value{
ggplot
}
\description{
Mosaic ggplot of xtabs object
}
\details{
Takes a table of exactly two categorical variables produced by xtabs and produces a ggplot mosaic plot
}
\examples{
\dontrun{
if(interactive()){
 ggxtabs(xtabs(~color+clarity, diamonds), viridis_option = 'H')
 }
}
}
