% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{inline_ttest}
\alias{inline_ttest}
\title{Output results of a t-test as a single string to be formatted inline by knitr}
\usage{
inline_ttest(ttest, effsize = NULL, sig = 3)
}
\arguments{
\item{effsize}{Include a user-supplied effect size (default = NULL)}

\item{sig}{The number of signficant digits for the output.}

\item{formula}{Specify t-test using standard formula interface.}

\item{data}{A data frame containing the variables in formula.}
}
\value{
A character vector with 1 element.
}
\description{
Output results of a t-test as a single string to be formatted inline by knitr
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ hp, mtcars)
m2 <- lm(mpg ~ wt, mtcars)
mnms <- c('MPG vs HP', 'MPG vs Weight')
vnms <- c('Miles per gallon'=mpg, 'Horsepower'=hp, 'Weight'=wt)
out <- forestplot(m1, m2, modelsnames=mnms, varnames=vnms)
}
