% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{joint_missingness}
\alias{joint_missingness}
\title{joint_missingness}
\usage{
joint_missingness(d, ...)
}
\arguments{
\item{d}{data frame}

\item{...}{Arguments passed on to upset function}
}
\value{
Upset plot
}
\description{
Upset plot of missinging by combinations of variables
}
\details{
Plot indicates number of missing cases for different combinations of variables
}
\examples{
\dontrun{
if(interactive()){
 joint_missingness(airquality)
 }
}
}
\seealso{
\code{\link[UpSetR]{upset}}
}
