% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{logistic_forestplot}
\alias{logistic_forestplot}
\title{logistic_forestplot}
\usage{
logistic_forestplot(
  ...,
  modelnames = NULL,
  varnames = NULL,
  stats = NULL,
  odds.ratio = F,
  breaks = NULL,
  intercept = T,
  sig = 3
)
}
\arguments{
\item{...}{One or more logistic regression model objects.}

\item{modelnames}{A character vector of names to label each plot and table column.}

\item{varnames}{A named character vector. Each element is a human friendly version of the variable name, and each name is the original variable name.}

\item{stats}{A named list of numeric vectors. The names are the names of the stats; the vectors are the stats to include in the table.}

\item{odds.ratio}{If TRUE, apply exponential transform to coefficients and omit intercept.}

\item{breaks}{An optional vector of breaks for the plot x-axis.}

\item{intercept}{If TRUE, include the intercept in the table and plot; otherwise omit.}

\item{sig}{The number of signficant digits for the table.}
}
\value{
A named list. $plot has the plot. $table has the data frame of coefficients and stats.
}
\description{
logistic_forestplot
}
\examples{
data(mtcars)
m1 <- glm(am ~ hp, family=binomial, mtcars)
m2 <- glm(mpg ~ wt, family=binomial, mtcars)
mnms <- c('AM vs HP', 'AM vs Weight')
vnms <- c(mpg='Miles per gallon', hp='Horsepower', wt='Weight')
out <- logistic_forestplot(m1, m2, modelsnames=mnms, varnames=vnms)
}
