% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{pca_biplot}
\alias{pca_biplot}
\title{pca_biplot}
\usage{
pca_biplot(
  obj,
  components = c(1, 2),
  data = NULL,
  threshold = 0,
  label_size = 5,
  geom_point = T
)
}
\arguments{
\item{obj}{prcomp obj}

\item{components}{The PCs to plot, specified as a numeric vector of length 2, Default: c(1, 2)}

\item{data}{Optional data frame with the same number of rows as that used to compute pca, Default: NULL}

\item{threshold}{Only display variables with loadings exceeding this value, Default: 0}

\item{label_size}{text size of the eigenvector labels, Default: 5}

\item{geom_point}{whether to add geom_point to plot, Default: T}
}
\value{
ggplot obj
}
\description{
Create a ggplot biplot of a prcomp obj
}
\details{
Creates a ggplot obj from a prcomp obj. Use threshold to omit variables with small loadings. Add a data frame to create a more complex ggplot. Possibly set geom_point=F to add a geom_point with custom aesthetics.
}
\examples{
\dontrun{
if(interactive()){
 m <- prcomp(mtcars, scale. = T)
 pca_biplot(m)
 pca_biplot(m, data = mtcars, geom_point=F) + geom_point(aes(size = mpg))
 }
}
}
