% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{pca_loadings_plot}
\alias{pca_loadings_plot}
\title{pca_loadings_plot}
\usage{
pca_loadings_plot(
  obj,
  components = 1:3,
  sortby = 1,
  threshold = 0,
  reverse = NULL
)
}
\arguments{
\item{obj}{Output of prcomp}

\item{components}{A numeric vector of components to plot}

\item{sortby}{Sort variables by loadings on this component (index to component vector)}

\item{threshold}{Remove variables with sum(abs(loadings)) below threshold (default=0)}

\item{reverse}{Vector of components that will be multiplied by -1}
}
\value{
ggplot object
}
\description{
A lollipop plot of PCA variable loadings,
as provided by prcomp, e.g., PC1, PC2, ...
}
\examples{
m <- prcomp(mtcars, scale = T)
pca_loadings_plot(m, components = c(1,2))
}
