% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{predicted_prob}
\alias{predicted_prob}
\title{Predicted probability from svyglm model and new data, along with 95\% CI.}
\usage{
predicted_prob(mod, df)
}
\arguments{
\item{mod}{A svyglm model object.}

\item{df}{A data frame of new data created with, e.g., expand.grid.}
}
\value{
df with three additional columns:
\describe{
\item{Probability}{The predicted probability at each value in df}
\item{LL}{The lower 95\% confidence limit}
\item{UL}{The upper 95\% confidence limit}
}
}
\description{
Predicted probability from svyglm model and new data, along with 95\% CI.
}
\examples{
m = glm(am ~ mpg + hp + cyl, family=binomial, data=mtcars)
newdata = expand.grid(mpg=10:35, hp=50:335, cyl=c(4,6,8))
df = predicted_prob(m, newdata)
}
