% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{regressiontable}
\alias{regressiontable}
\title{regressiontable}
\usage{
regressiontable(models, caption = NULL, sigfig = 3)
}
\arguments{
\item{models}{Named list of lm and glm models.}

\item{caption}{Optional caption for table, Default: NULL}

\item{sigfig}{The number of significant digits to display, Default: 3}
}
\value{
A gt table
}
\description{
Create gt table of regression coefficients
}
\details{
Currently only supports lm and glm models
}
\examples{
\dontrun{
if(interactive()){
 models <- list(
 "model 1" = lm(mpg ~ hp, mtcars),
 "model 2" = glm(am ~ hp, family=binomial, mtcars)
 )
 regressiontable(models)
 }
}
}
\seealso{
\code{\link[broom]{tidy}}
 \code{\link[gt]{gt}}
}
