% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{renumber_tweets}
\alias{renumber_tweets}
\title{renumber_tweets}
\usage{
renumber_tweets(fn)
}
\arguments{
\item{fn}{file name of markdown file}
}
\value{
Writes to the input file, fn!
}
\description{
Replace #. with sequential numbering (e.g., 1., 2.,)
}
\details{
Use when composing twitter thread in markdown. Number tweets with #. placeholder.
}
\examples{
\dontrun{
if(interactive()){
 renumber_tweets("thread.md")
 }
}
}
