% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{reorder_levels}
\alias{reorder_levels}
\title{reorder_levels}
\usage{
reorder_levels(f, v, index, choice)
}
\arguments{
\item{f}{Character vector used as a factor variable}

\item{v}{Numeric vector used to order the factor levels}

\item{index}{Vector used to divide f and v into exactly 2 groups (e.g., Males, Females)}

\item{choice}{The value of index to use to (e.g., Female)}
}
\value{
An character vector to be used in the level argument of factor()
}
\description{
Order factor levels based on a subset of values (e.g., only female values, or only male values)
}
