% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{scale2}
\alias{scale2}
\title{scale2}
\usage{
scale2(x)
}
\arguments{
\item{x}{A numeric matrix(like object).}
}
\value{
A scaled vector or matrix or array
}
\description{
For numeric vectors, returns a scaled numeric vector (not an array); otherwise returns scale(x)
}
\examples{
x <- scale2(mtcars$mpg)
class(x)
}
