% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{scale_colour_binary}
\alias{scale_colour_binary}
\alias{scale_color_binary}
\title{scale_colour_binary}
\usage{
scale_colour_binary(direction = 1, ...)

scale_color_binary(direction = 1, ...)
}
\arguments{
\item{direction}{Order of colors, 1: purple-red. -1: red-purple, Default: 1}

\item{...}{Additional arguments to be passed to ggplot2::discrete_scale}
}
\value{
a colour scale to be used with ggplot2
}
\description{
Provides exactly two colors from the viridis magma palette (red and purple)
}
\details{
The default top level from the discrete viridis scale is bright yellow, which is too bright. This scale uses red for top level.
}
\examples{
\dontrun{
if(interactive()){
 ggplot(mtcars, aes(hp, mpg, colour=factor(am))) + geom_point() + scale_colour_binary()
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
 \code{\link[viridisLite]{magma}}
}
