% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{scale_fill_binary}
\alias{scale_fill_binary}
\title{scale_fill_binary}
\usage{
scale_fill_binary(direction = 1, ...)
}
\arguments{
\item{direction}{Order of colors, 1: purple-red. -1: red-purple, Default: 1}

\item{...}{Additional arguments to be passed to ggplot2::discrete_scale}
}
\value{
a fill scale to be used with ggplot2
}
\description{
Provides exactly two colors from the viridis magma palette (red and purple)
}
\details{
The default top level from the discrete viridis scale is bright yellow, which is too bright. This scale uses red for top level.
}
\examples{
\dontrun{
if(interactive()){
 ggplot(mtcars, aes(mpg, fill=factor(am))) + geom_histogram() + scale_fill_binary()
 }
}
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
 \code{\link[viridisLite]{magma}}
}
