% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{signif2}
\alias{signif2}
\title{signif2}
\usage{
signif2(x, sigfigs = 2, format = "g")
}
\arguments{
\item{x}{Numeric vector}

\item{sigfigs}{Number of significant digits, Default: 2}

\item{format}{'g': use scientific notation when more compact; 'fg': numbers in xx.yy format, Default: 'g'}
}
\value{
Character vector of numbers formatted to the specified number of sigfigs
}
\description{
Convert numeric vector to character vector with specified number of significant digits, retaining trailing zeros
}
\details{
Character and logical vectors are returned as is.
}
\examples{
\dontrun{
if(interactive()){
 signif2(rnorm(10))
 }
}
}
