% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{summaryTable}
\alias{summaryTable}
\title{summaryTable}
\usage{
summaryTable(d, vars = NULL, fac = NULL, statscols = T, sigfigs = 2)
}
\arguments{
\item{d}{Data frame}

\item{vars}{Named vector. Names are variables to summarize, values are human readable versions., Default: NULL}

\item{fac}{Character value of grouping variable, Default: NULL}

\item{statscols}{Include Cohen's d and a Mann-Whitney test if a fac is provided, Default: T}

\item{sigfigs}{Number of significant figures to report, Default: 2}
}
\value{
Either a single kable object, or a list of two kable objects
}
\description{
Provides summary stats for a data frame, possibly by a two-level factor
}
\examples{
\dontrun{
if(interactive()){
 summaryTable(mtcars, fac="am", vars = c(mpg="Miles per gallon", hp="Horsepower"))
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
 \code{\link[tibble]{tibble}}
 \code{\link[glue]{glue}}
 \code{\link[purrr]{map}}
 \code{\link[knitr]{kable}}
 \code{\link[kableExtra]{add_header_above}}
}
