% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{svysmooth2df}
\alias{svysmooth2df}
\title{svysmooth2df}
\usage{
svysmooth2df(...)
}
\arguments{
\item{...}{One or more svysmooth objects with names.}
}
\value{
A data frame (tibble) with 3 columns: x, y, and a
character vector indicating which values belong to which smooth
}
\description{
Extracts the x and y values from one or more svysmooth objects,
from the survey package, and returns them in a data frame.
}
\details{
If multiple smooths are supplied, the x and y variables
must be the same, but computed over different design objects (e.g.,
different subsets of a survey).
}
\examples{
library(survey)
data(api)
dclus1<-svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
dclus2<-subset(dclus1, sch.wide=="Yes" & comp.imp=="Yes")
m1 <- svysmooth(api00~ell,dclus1)
m2 <- svysmooth(api00~ell,dclus2)
df <- svysmooth2df(full = m1, subset = m2)
}
