% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{tjurD}
\alias{tjurD}
\title{tjurD}
\usage{
tjurD(m)
}
\arguments{
\item{m}{A glm object with family = binomial}
}
\value{
A list with two elements, rsq = Tjur's D, and plot = ggplot object.
}
\description{
Computes Tjur's D
}
\details{
Computes Tjur's D and produces a probability plot
}
\examples{
\dontrun{
if(interactive()){
 m <- glm(am ~ mpg, family = binomial, mtcars)
 out <- tjurD(m)
 out$rsq
 out$plot
 }
}
}
