% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{twittersize}
\alias{twittersize}
\title{twittersize}
\usage{
twittersize(
  path,
  aspect_ratio = "16:9",
  out_prefix = "twitter ",
  border_color = "white"
)
}
\arguments{
\item{path}{Path to image file}

\item{aspect_ratio}{Desired aspect ratio, Default = "16:9" (single image in tweet). Use "7:8" if tweeting two images.}

\item{out_prefix}{Prepend to name of output image, Default = "twitter"}

\item{border_color}{Border color, Default = "white"}
}
\value{
path to output image
}
\description{
Add borders to an image to force an aspect ratio for twitter post, and write it out in png format. Does not resize original image.
}
\examples{
\dontrun{
if(interactive()){
 twittersize('~/Desktop/image.png')
 }
}
}
