% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{cohen_d}
\alias{cohen_d}
\title{Cohen's d}
\usage{
cohen_d(f, data, sig = 3)
}
\arguments{
\item{f}{A two-sided formula: y ~ x.}

\item{d}{A data frame.}
}
\value{
Cohen's d
}
\description{
Cohen's d
}
\examples{
cohen_d(mpg ~ am, mtcars)
}
