% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{fmt_terms}
\alias{fmt_terms}
\title{fmt_terms}
\usage{
fmt_terms(models)
}
\arguments{
\item{models}{A named list of models}
}
\value{
A name list of model statistics, including str, a string version of the term estimate.
}
\description{
Format estimate and 95\% CIs for inline markdown
}
\details{
Provide a named list of regression models, and the resulting named list will have a formatted string suitable for inline rmarkdown
}
\examples{
\dontrun{
if(interactive()){
 stats <- fmt_terms(list(m1 = lm(mpg ~ hp, mtcars), m2 = lm(mpg ~ wt, mtcars)))
 stats$m1$hp$str
 }
}
}
