% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ggemmeans}
\alias{ggemmeans}
\title{ggemmeans}
\usage{
ggemmeans(em)
}
\arguments{
\item{em}{An emmeans object}
}
\value{
a ggplot object
}
\description{
ggplot of estimated marginal means object, with categorical spec variable sorted by estimate
}
\details{
This function is similar to the emmeans plot function, except that it sorts the levels of a categorical variable by the estimated mean
}
\examples{
\dontrun{
if(interactive()){
 m <- lm(height ~ eye_color, starwars)
 em <- emmeans(m, spec = 'eye_color')
 ggemmeans(em)
 }
}
}
