% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{ggxtabs}
\alias{ggxtabs}
\title{ggxtabs}
\usage{
ggxtabs(xtab, cell_counts = F, viridis_option = "B")
}
\arguments{
\item{xtab}{xtabs object of exactly two categorical variables}

\item{cell_counts}{Display cell counts, Default: F}

\item{viridis_option}{Viridis color palette, Default: 'B'}
}
\value{
ggplot
}
\description{
Mosaic ggplot of xtabs object
}
\details{
Takes a table of exactly two categorical variables produced by xtabs and produces a ggplot mosaic plot
}
\examples{
\dontrun{
if(interactive()){
 ggxtabs(xtabs(~color+clarity, diamonds), viridis_option = 'H')
 }
}
}
