% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{hagenheat}
\alias{hagenheat}
\title{hagenheat}
\usage{
hagenheat(
  d,
  method = "seriate",
  seriation_method = NULL,
  independent = F,
  hc_method = "ward.D",
  dist_method = "euclidean",
  scale. = "none",
  viridis_option = "D",
  wrap_labels = T,
  rotate_labels = F,
  ann_col = NULL
)
}
\arguments{
\item{d}{A data frame, matrix, or dist object. Column 1 of a data frame can be row labels. Remaining columns must be numeric.}

\item{method}{"seriate" or "hclust". Default: "seriate"}

\item{seriation_method}{If not specified, 'Spectral' will be used for dist objects, and 'PCA' for matrices and data frames.}

\item{independent}{Whether to seriate a matrix using matrix methods (F) or separate distance matrices for rows and columns (T). Default: F}

\item{hc_method}{Agglomeration method from hclust, Default: 'ward.D'}

\item{scale.}{Whether to scale rows ('row'), columns ('column'), or neither ('none'), Default: 'none'}

\item{viridis_option}{One of the viridis color options, 'A', 'B', 'C', 'D', 'E', Default: 'D'}

\item{wrap_labels}{Logical. Wrap the x-axis labels. Default: T}

\item{rotate_labels}{Logical. Rotate the x-axis labels by 90 degrees. Default: F}

\item{ann_col}{A data frame with two variables. The first column must be the names of the columns of 'd'. The second column contains the annotation values.}

\item{dist}{Distance method from dist, Default: 'euclidean'}
}
\value{
A ggplot object
}
\description{
Basic heatmap using ggplot, dist, and seriation. No dendrograms are plotted.
}
\details{
Produces a very simple ggplot heatmap using viridis colors. 
Rows and columns are ordered using the \code{seriation} package.
For data frames, first column must be row labels. Remaining columns must be numeric.
Scaling, if performed, occurs after seriation, and is therefore only for display purposes.
Seriation methods for objects when \code{independent=T}: \code{list_seriation_methods('dist')}.
Seriation methods for matrices and data frames when \code{independent=F}: \code{list_seriation_methods('matrix')}.
}
\examples{
\dontrun{
if(interactive()){
 hagenheat(mtcars, scale. = 'col')
 }
}
}
\seealso{
\code{\link[seriation]{seriate}}
 \code{\link[seriation]{seriation_methods}}
 \code{\link[ggnewscale]{new_scale_fill}}
}
