% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{hagenheat}
\alias{hagenheat}
\title{hagenheat}
\usage{
hagenheat(
  d,
  method = "seriate",
  seriation_method = "Spectral",
  hc_method = "ward.D",
  dist_method = "euclidean",
  scale. = "none",
  viridis_option = "D",
  ann_col = NULL
)
}
\arguments{
\item{d}{A data frame, matrix, or dist object. Column 1 of a data frame can be row labels. Remaining columns must be numeric.}

\item{method}{"seriate" or "hclust". Default: "seriate"}

\item{seriation_method}{Default 'Spectral'}

\item{hc_method}{Agglomeration method from hclust, Default: 'ward.D'}

\item{scale.}{Whether to scale rows ('row'), columns ('column'), or neither ('none'), Default: 'none'}

\item{viridis_option}{One of the viridis color options, 'A', 'B', 'C', 'D', 'E', Default: 'D'}

\item{ann_col}{A data frame with two variables. The first column must be the names of the columns of 'd'. The second column contains the annotation values.}

\item{dist}{Distance method from dist, Default: 'euclidean'}
}
\value{
A ggplot object
}
\description{
Basic heatmap using ggplot, dist, and hclust. No dendrograms are plotted, though.
}
\details{
Produces a very simple ggplot heatmap using viridis colors. 
Rows and columns are clustered using the seriation package.
First column of data frame must be row labels. Remaining columns must be numeric.
}
\examples{
\dontrun{
if(interactive()){
 hagenheat(mtcars, scale. = 'col')
 }
}
}
\seealso{
\code{\link[seriation]{seriate}}
 \code{\link[seriation]{seriation_methods}}
 \code{\link[viridis]{scale_color_viridis}}
 \code{\link[ggnewscale]{new_scale_fill}}
}
