% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{hagenheat}
\alias{hagenheat}
\title{hagenheat}
\usage{
hagenheat(d, hc_method = "ward.D", dist = "euclidean", scale. = "row")
}
\arguments{
\item{d}{A data frame. Column 1 must be row labels. Remaining columns must be numeric.}

\item{hc_method}{Agglomeration method from hclust, Default: 'ward.D'}

\item{dist}{Distance method from dist, Default: 'euclidean'}

\item{scale.}{Whether to scale rows ('row'), columns ('col'), or neither ('none'), Default: 'row'}
}
\value{
A ggplot object
}
\description{
Basic heatmap using ggplot. No dendrograms, though.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 d <- cbind(rownames(mtcars), mtcars)
 hagenheat(d, scale. = 'col')
 }
}
}
